+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # ChangeLog
 
+## 0.2.0 - March 19, 2014
+* Fixed `Identity` using old import.
+* `Identity` as functor.
+* `ContT` and `Cont` as functor.
+* `DContT` and `DCont` as functor.
+* `ListT` and `List` as functor.
+* `StateT` and `State` as functor.
+
 ## 0.1.0 - March 19, 2014
 * Initial release.